Update pharmaceuticalLogs.c
This is a computer science algorithms project I made during by bachelors of Computer Science. This project incorporates multiple algorithms such as the Merge Sort, the Binary Search, the Linear Search, initialization of a linked list, adding a node to a linked list and reading from a linked list. I designed this project so each algorithm is kept within a certain efficiency or better.